fix(frontend): fix dashboard crash, rename to kebab-case, fix error patterns
